* include/math.h (nanl, __fpcassifyl, fminl, fmaxl, rintl, 	roundl, truncl, fmal, log2l): Declare. 	Protect C99 declarations with _STDC_VERSION__ >= 199901L) 	|| !defined  __STRICT_ANSI__.  	* mingwex/fmax.c (fmax): Call __isnan, not _isnan. 	* mingwex/fmin.c (fmin): Likewise. 	* mingwex/fmaxf.c (fmaxf): Call __isnanf, not _isnan. 	* mingwex/fminf.c (fminf): Likewise. 	* mingwex/fmaxl.c: New file. 	* mingwex/fminl.c: New file. 	* mingwex/fpclassify.c (__fpclassifyf): Split out to ... 	* mingwex/fpclassifyf.c: New file. 	* mingwex/fpclassifyl.c: New file. 	* mingwex/rint.c (rintf): Split out to... 	* mingwex/rintf.c: New file. 	* mingwex/rintl.c: New file. 	* mingwex/round.c (roundf): Split out to... 	* mingwex/roundf.c: New file. 	* mingwex/roundl.c: New file. 	* mingwex/trunc.c (truncf): Split out to... 	* mingwex/truncf.c: New file. 	* mingwex/truncl.c: New file. 	* mingwex/signbit.c (signbitf): Split out to... 	* mingwex/signbitf.c: New file. 	* mingwex/signbitl.c: New file. 	* mingwex/fmal.c: New file. 	* mingwex/copysignl.S: New file. 	* mingwex/log2l.c: New file. 	* mingwex/fp_consts.c: Add nanl definition. 	Comment out unused constants. 	* mingwex/Makefile.in (DISTFILES): Add fmaxl.c, fminl.c, 	fpclassifyf.c, fpclassifyl.c, rintf.c, rintl.c, roundf.c, 	roundl.c, truncf.c truncl.c, signbitf.c signbitl.c, 	fmal.c, copysignl.S, log2l.c 	(MATHOBJS): Add fmaxl.o, fminl.o, fpclassifyf.o, 	fpclassifyl.o, rintf.o, rintl.o, roundf.o, roundl.o, 	truncf.o truncl.o, signbitf.o signbitl.o, fmal.o, 	copysignl.o, log2l.o.

+++ winsup/mingw/mingwex/fucom.c
@@ -1,5 +1,5 @@
 int 
-__fp_unordered_compare (double x,  double y){
+__fp_unordered_compare (long double x,  long double y){
   unsigned short retval;
   __asm__ (
 	"fucom %%st(1);"
